+++ core/src/main/java/org/elasticsearch/http/HttpServerAdapter.java
@@ -20,11 +20,13 @@
 package org.elasticsearch.http;
 
 import org.elasticsearch.common.util.concurrent.ThreadContext;
+import org.elasticsearch.rest.RestChannel;
+import org.elasticsearch.rest.RestRequest;
 
 /**
  *
  */
 public interface HttpServerAdapter {
 
-    void dispatchRequest(HttpRequest request, HttpChannel channel, ThreadContext context);
+    void dispatchRequest(RestRequest request, RestChannel channel, ThreadContext context);
 }
